… cmd/dicomtest/main.go