@suyashkumar 2020-07-17
btw, query check of parse performance of a huge dicom is about the same for both old and new, a good sign (new, a teeny bit faster):

Old:
`./build/dicomutil --extract-images-stream <REDACTED PATH>   23.58s user 2.47s system 178% cpu 14.611 total`

New:
`./dcm -path <REDACTED PATH>   22.86s user 2.29s system 186% cpu 13.510 total`